acpioff: Force a human to reboot system, if turning off power fails
Once the hardare has been put in ACPI mode, one must assume that it can
never reliably be reverted back to the start-up state.  This can be a
source of problems for any ACPI-aware OS or COM32 module loaded after
acpioff, if acpioff fails to turn off power.  Force a human to reboot the
system, before any other image can be subsequently loaded and run.

Thanks go to H. Peter Anvin and Gene Cumm for pointing this out.

Signed-off-by: Andy Walls <awalls@md.metrocast.net>